chore: refactoring fastapi models
